Fix: Move useEffect that checks if user wants to leave page to audioPlayer component
