+++ src/App.jsx
@@ -70,20 +70,6 @@ function App() {
         setIsResetDialogOpen(false);
     };   
 
-    useEffect(() => {
-        const handleBeforeUnload = (event) => {
-          if (isPlaying) {
-            event.preventDefault();
-            event.returnValue = ''; // This is required for some browsers
-          }
-        };
-    
-        window.addEventListener('beforeunload', handleBeforeUnload);
-    
-        return () => {
-          window.removeEventListener('beforeunload', handleBeforeUnload);
-        };
-      }, [isPlaying]);
   
     // const handleSkipNext = () => {
     //     if (detailedShow && currentEpisode) {
